Migrate JobsAPI to kamiwaza_sdk/services/jobs_federation.py (T7.6 / ENG-5040)
WS-M3.2 mechanical service migration. The M1+ JobsAPI already used
self._client._request(method, path, json=body), so the relocation is a
pure file move + import switch. Class inherits from BaseService;
references self.client instead of self._client.

Renamed to jobs_federation.py per design §6.2 T7.6 to leave room for a
future legacy-style local-cluster jobs service. Class name JobsAPI
preserved so existing M1-M3 callsites work unchanged via re-export.

Legacy kamiwaza/jobs.py becomes a thin re-export. 11 existing
test_kamiwaza_jobs_*.py tests pass via the re-export — identity
invariant holds.

References: ENG-5040 (T7.6), design v0.3.7 §4.2.11.
